- renaming all references for flux to redux
- delete unused test file
