+++ app/build.gradle
@@ -5,7 +5,7 @@ android {
     compileSdkVersion 24
     buildToolsVersion "24.0.3"
     defaultConfig {
-        applicationId "br.com.catbag.giffluxsample"
+        applicationId "br.com.catbag.gifreduxsample"
         minSdkVersion 21
         targetSdkVersion 24
         versionCode 1
