+++ app/src/main/AndroidManifest.xml
@@ -1,18 +1,18 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="br.com.catbag.giffluxsample">
+          package="br.com.catbag.gifreduxsample">
 
     <uses-permission android:name="android.permission.INTERNET"/>
     <uses-permission android:name="android.permission.SET_DEBUG_APP"></uses-permission>
 
     <application
-        android:name=".App"
+        android:name="br.com.catbag.gifreduxsample.App"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".ui.GifListActivity">
+        <activity android:name="br.com.catbag.gifreduxsample.ui.GifListActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
