+++ app/src/main/java/br/com/catbag/gifreduxsample/App.java
@@ -1,12 +1,12 @@
-package br.com.catbag.giffluxsample;
+package br.com.catbag.gifreduxsample;
 
 import android.app.Application;
 
 import com.umaplay.fluxxan.Fluxxan;
 
-import br.com.catbag.giffluxsample.models.AppState;
-import br.com.catbag.giffluxsample.models.ImmutableAppState;
-import br.com.catbag.giffluxsample.reducers.GifReducer;
+import br.com.catbag.gifreduxsample.models.AppState;
+import br.com.catbag.gifreduxsample.models.ImmutableAppState;
+import br.com.catbag.gifreduxsample.reducers.GifReducer;
 
 /**
  * Created by niltonvasques on 10/12/16.
