+++ app/src/main/java/br/com/catbag/gifreduxsample/asyncs/restservice/ApiRoutes.java
@@ -1,4 +1,4 @@
-package br.com.catbag.giffluxsample.asyncs.restservice;
+package br.com.catbag.gifreduxsample.asyncs.restservice;
 
 import okhttp3.ResponseBody;
 import retrofit2.Call;
