+++ app/src/main/res/layout/activity_gif_list.xml
@@ -9,7 +9,7 @@
                 android:paddingRight="@dimen/activity_horizontal_margin"
                 android:paddingTop="@dimen/activity_vertical_margin"
                 android:background="@color/notWatched"
-                tools:context="br.com.catbag.giffluxsample.ui.GifListActivity">
+                tools:context="br.com.catbag.gifreduxsample.ui.GifListActivity">
 
     <ImageView
         android:id="@+id/gif_image"
