+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Gif Flux Sample</string>
+    <string name="app_name">Gif Redux Sample</string>
 </resources>
